emacsへの$PATHの書き出し・npm completionの読み込み・make cleanを安全に

+++ Makefile
@@ -2,12 +2,16 @@ BASH_CONF = ~/bash.conf
 
 all: ~/.bashrc ~/.bash_profile
 
-~/.bashrc:
+~/.bashrc: ~/.npm-completion
 	ln -s $(BASH_CONF)/bashrc.sh ~/.bashrc
 
 ~/.bash_profile:
 	ln -s $(BASH_CONF)/bashrc.sh ~/.bash_profile
 
+~/.npm-completion:
+	npm completion > $(BASH_CONF)/.npm-completion
+
 clean:
-	rm ~/.bashrc
-	rm ~/.bash_profile
+	if [ -L ~/.bashrc ]; then rm ~/.bashrc ; fi
+	if [ -L ~/.bash_profile ]; then rm ~/.bash_profile ; fi
+	if [ -L ~/.npm-completion ]; then rm ~/.npm-completion ; fi
